Fix error delete study mistake huytd'

+++ app/Http/Controllers/Api1/Studies.php
@@ -147,16 +147,13 @@ class Studies extends Controller
 
     public function deleteStudyById($id)
     {
-        /** @var Study $study */
         $study = Study::findOrFail($id);
         if (!$study) return -1;
         
         $conf = $this->kernel->getConfig($this->auth->user()->ID_USER, intval($id), -1);
         $this->kernel->getKernelObject('StudyCleaner')->SCStudyClean($conf, SC_CLEAN_OUTPUT_ALL);
         
-        /** @var Product[] $product */
         $products = $study->products;
-        
         foreach ($products as $product) {
             // 3d featrue delete mesh3D_info
             $mesh3D_info = Mesh3DInfo::Where('ID_PROD', $product->ID_PROD)->first();
@@ -175,7 +172,6 @@ class Studies extends Controller
                 $mesh3D_info->delete();
             }
 
-            /** @var MeshGeneration $meshGenerations */
             $meshGenerations = $product->meshGenerations;
             foreach ($meshGenerations as $mesh) {
                 $mesh->delete();
@@ -183,9 +179,12 @@ class Studies extends Controller
 
             foreach ($product->productElmts as $productElmt) {
                 // 3d featrue not delete mesh position
-                // foreach ($productElmt->meshPositions as $meshPst) {
-                //     $meshPst->delete();
-                // }
+                foreach ($productElmt->meshPositions as $meshPst) {
+                    if ($meshPst) {
+                        $meshPst->delete();
+                    }
+                }
+
                 // 3d feature delete initial_temp
                 $initial3D_temps = InitTemp3D::where('ID_PRODUCT_ELMT', $productElmt->ID_PRODUCT_ELMT)->get();
                 if (count($initial3D_temps) > 0) {
@@ -196,8 +195,10 @@ class Studies extends Controller
                 $productElmt->delete();
             }
 
-            foreach ($product->prodcharColors as $prodCharColor) {
-                $prodCharColor->delete();
+            if (count($product->prodcharColors) > 0) {
+                foreach ($product->prodcharColors as $prodCharColor) {
+                    $prodCharColor->delete();
+                }
             }
 
             $product->delete();
@@ -208,14 +209,16 @@ class Studies extends Controller
         foreach ($study->prices as $price) {
             $price->delete();
         }
+
         // 3d featrue not delete Initial_temperature
-        // foreach ($productions as $production) {
-        //     InitialTemperature::where('ID_PRODUCTION', $production->ID_PRODUCTION)->delete();
-        //     $production->delete();
-        // }
+        if (count($productions) > 0) {
+            foreach ($productions as $production) {
+                InitialTemperature::where('ID_PRODUCTION', $production->ID_PRODUCTION)->delete();
+                $production->delete();
+            }
+        }
 
         $tempRecordPts = $study->tempRecordPts;
-
         foreach ($tempRecordPts as $tempRecord) {
             $tempRecord->delete();
         }
@@ -224,7 +227,6 @@ class Studies extends Controller
             $report->delete();
         }
         
-        /** @var Packing $packing */
         foreach ($study->packings as $packing) {
             foreach ($packing->packingLayers as $packingLayer) {
                 $packingLayer->delete();
